requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
Click==7.0
Django==2.1.5
djangorestframework==3.9.0
pytz==2018.7
toml==0.10.0
flake8
psycopg2>=2.7.3.2
